+++ src/main/java/com/eoi/Fruggy/repositorios/RepoTipoDescuento.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepoTipoDescuento extends JpaRepository<TipoDescuento, Integer> {
+public interface RepoTipoDescuento extends JpaRepository<TipoDescuento, Long> {
 }
